requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
executing==2.2.1
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jiter==0.12.0
json_repair==0.54.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.16
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.64
langchain-ollama==0.1.3
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.26.1
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
neo4j==6.0.3
numpy==1.26.4
ollama==0.6.1
openai==1.109.1
orjson==3.11.5
packaging==24.2
parso==0.8.5
prompt_toolkit==3.0.52
propcache==0.4.1
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.4.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.45
stack-data==0.6.3
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
widgetsnbextension==4.0.15
yarl==1.22.0
yfiles_jupyter_graphs==1.10.9
